services/alert-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
redis==5.0.4
apscheduler==3.10.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
ruff==0.4.4
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
